Verify and optimize src/ against .claude/rules conventions
- Fix undefined $commandName / $parameterBag properties in console commands (PHPUnit, Test, CloudflareR2, I18n)
- Fix ComposerCommand: null returns from execute(), call to undefined outputWithTime(), finally-swallowed throws with unbounded retry recursion (now bounded to one retry)
- Fix LazyEntityCommand undefined $className; AuroraChronos dead instanceof checks, string return from daysBetweenTwoDates() and str_pad() int arguments; AuroraGit date() string timestamp
- Fix PHP 8.4 implicitly-nullable parameter in CustomExceptionController; int/float return mismatches in AbstractTimestampableDeleted* lifespan methods
- Replace trigger_error(E_USER_DEPRECATED) with trigger_deprecation(); replace backticks in exception messages with sprintf() and double quotes
- Initialize compressCSSJS() output variables unconditionally; type private properties and badge method parameters
- Remove dead code: buggy cssMinify() v1 branch, PHP 5.4 version check in HtmlFormatter, redundant @var blocks, dead null checks
- PHPStan level 6 on src/: 546 -> 463 errors; all direct tests of the modified classes pass (108/108)

+++ src/Entity/SuperClass/AbstractTimestampableDeletedNullable.php
@@ -63,18 +63,18 @@ abstract class AbstractTimestampableDeletedNullable
     #[Groups([AuroraConstants::GROUP_READ])]
     public function getDeletedAtLifespanAsMinutes(): int
     {
-        return round($this->getDeletedAtLifespanAsSeconds() / 60);
+        return (int)round($this->getDeletedAtLifespanAsSeconds() / 60);
     }
 
     #[Groups([AuroraConstants::GROUP_READ])]
     public function getDeletedAtLifespanAsHours(): int
     {
-        return round($this->getDeletedAtLifespanAsMinutes() / 60);
+        return (int)round($this->getDeletedAtLifespanAsMinutes() / 60);
     }
 
     #[Groups([AuroraConstants::GROUP_READ])]
     public function getDeletedAtLifespanAsDays(): int
     {
-        return round($this->getDeletedAtLifespanAsHours() / 24);
+        return (int)round($this->getDeletedAtLifespanAsHours() / 24);
     }
 }
